+++ src/chat/websocket/sendChatMessage.ts
@@ -1,6 +1,7 @@
 import { SendMessagePayload } from "../types/websocket";
-import { stompClient } from "./socketClient";
+import { stompClient } from "./connectStomp";
 
+// 채팅 메세지 웹소켓으로 전송
 export const sendChatMessage = (payload: SendMessagePayload) => {
     try {
         if (stompClient.connected) {
